documentgenerator-foxtrot-mh/requirements.txt
python-dotenv
openai
azure-ai-documentintelligence
python-pptx
python-docx
streamlit==1.43.2